Always release dhcp address in close_tun()
Also make sure --dhcp-pre-release results in not just dhcp_release()
in open_tun() but a subsequent dhcp_renew() as well. Else dhcp transaction
gets aborted as this call to release() happens after the adapter status
is changed to connected.

Fixes Trac #807 (but can't say the same for Trac #665 without knowing
how to reproduce it)

v2: Mark --dhcp-release as obsolete in mapage and option parser, and
remove the unused dhcp_release varaible.
Enforce dhcp-renew with dhcp-pre-release while parsing the option
instead of in open_tun().

Signed-off-by: Selva Nair <selva.nair@gmail.com>