Fix user's group membership check in interactive service to work with domains
Currently the username unqualified by the domain is used to validate
a user which fails for domain users. Instead

- the user is authorized if the admin or ovpn_admin group is in the token
- else if the user's SID is in the admin or ovpn_admin group

The second check is needed to recognize dynamic updates to group membership
on the local machine.

These checks do no require connection to a domain controller and will
work even when user is logged in with cached credentials.

Resolves Trac: #810

v2: include the token check as described above

Signed-off-by: Selva Nair <selva.nair@gmail.com>