Refactor parsing of --iroute option
Functions option_iroute() and option_iroute_ipv6() now returns
falsevon error and are made non-static.

The only functionality change is that --iroute target can
now be specified in CIDR notation (like --iroute-ipv6) in
addition to the currently allowed "network netmask" notation:

   --iroute network[/bits] [netmask]

If both /bits and netmask are specified, the netmask is
ignored and a message is logged.

This refactoring helps the proposed patch for adding iroute
on the fly through the mangement interface.

Signed-off-by: Selva Nair <selva.nair@gmail.com>