Support adding internal route to a connected client from management
Implement a new management command

   --add-iroute network/bits destination

This adds an internal route to target specified as network/bits
via client specified by "destination". The latter could be
the VPN ifconfig_local IP of the client (IPv4 or IPv6), the client-id
(CID) of the client or a string "CN:common-name" obtained by prepending
"CN:" to the common-name of the client.

Note that the destination address is not the real-ip of the client
unlike in kill-by-addr command.

The target network must be an IPv4 or IPv6 network address with
bits specifying the prefix length. If /bits is omitted, it defaults
to 32 for IPv4 or 64 for IPv6 matchig with the --iroute and
--iroute-ipv6 options in the client-specific config files.

Examples:
   add-iroute 192.168.5.0/24 21          # client by CID
   add-iroute 2001::0b:/64 2001::0a:1005 # client by IPv6
   add-iroute 192.168.10.11 10.11.0.6    # host route, client by IPv4

Notes:
(i) Client-id as destination is available only if the daemon is built with
MANAGEMENT_DEF_AUTH support (this is the default unless --disable-def-auth
is used while configuring)

(ii) As the common-name need not be unique, the first client that matches
the specified name is used

(iii) These routes do not persist across reconnections. For persistent
iroutes also add a matching iroute directive to the ccd file for the client.

(iv) The route for a client is added only if the client has completed
authentication, ccd-file parsing and client-connect-script, if any.

Signed-off-by: Selva Nair <selva.nair@gmail.com>