travis: fix CFLAGS assignment error and add -Werror only when compiling on Linux for Linux
CFLAGS may not resolve to anything and the build-check.sh script would
just fail.

Switch to EXTRA_FLAGS and use proper substitution so that the script won't
error out when EXTRA_FLAGS is uninitialized.

At the same time ensure that such flag is set only when compiling *for*
linux (we use linux to also compile *for* windows)

Signed-off-by: Antonio Quartulli <a@unstable.cc>