Improve signal handling using POSIX sigaction
Currently we use the old signal API which follows system-V or
BSD semantics depending on the platform and/or feature-set macros.
Further, signal has many weaknesses which makes proper masking
(blocking) of signals during update not possible.

Improve this:

- Use sigaction to properly mask signals when modifying.

- Change signal_reset() to read the current value and reset in one
  operation.  This is required to avoid change of signal state between
  check and reset-operations.
  This also allows us to eliminate resetting signal to 0 in
  init_instance() which can potentially lose signals. Instead, the signal
  is reset at the end of the SIGUSR1 and SIGHUP loops where their
  values are checked.

Notes:

SIG_SOURCE_CONNECTION_FAILED is retained in a hackish way. This value
has the same meaning as SIG_SOURCE_SOFT everywhere except where the
signal is printed. Looks cosmetic --- could be eliminated?

SIGUSR1 during dns resolution is ignored and reset to zero in the original
and that behaviour is retained. Not sure why this is needed.
Special handling of signals in openvpn_getaddrinfo() as if it is a syscall
is retained but looks superfluous.

In pre_init_signal_catch() we ignore some unix signals, but the same signals
from management are not ignored though both are treated as "HARD" signals.
For example, during auth-user-pass query, "kill -SIGUSR1 <pid>" will be ignored,
but "signal SIGUSR1" from management interface will cause M_FATAL and exit.
This is the current behaviour, but could be improved?

v2: A buggy version was submitted as v1 -- NULL check missing around line 1785 in socket.c
    Fixed.

Signed-off-by: Selva Nair <selva.nair@gmail.com>